ref: refactor out rpc calls ui, authprovider, etc
